Backend/application-de-gestion-de-salle-de-sport/backend/requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
psycopg2-binary
python-dotenv
djoser
django-filter
django-rest-knox
celery
redis
django-celery-beat
django-celery-results
pytest
pytest-django
coverage
black
flake8
isort
